+++ cpukit/score/src/objectallocate.c
@@ -73,7 +73,7 @@ Objects_Control *_Objects_Allocate(
     if ( the_object ) {
       uint32_t   block;
 
-      block = _Objects_Get_index( the_object->id ) -
+      block = (uint32_t) _Objects_Get_index( the_object->id ) -
               _Objects_Get_index( information->minimum_id );
       block /= information->allocation_size;
