+++ cpukit/score/src/objectfree.c
@@ -50,8 +50,8 @@ void _Objects_Free(
   if ( information->auto_extend ) {
     uint32_t    block;
 
-    block =
-      _Objects_Get_index( the_object->id ) - _Objects_Get_index( information->minimum_id );
+    block = (uint32_t) (_Objects_Get_index( the_object->id ) -
+                        _Objects_Get_index( information->minimum_id ));
     block /= information->allocation_size;
 
     information->inactive_per_block[ block ]++;
